+++ cfgov/unprocessed/js/molecules/GlobalBanner.js
@@ -26,7 +26,7 @@ function GlobalBanner( element ) {
   var BASE_CLASS = 'm-global-banner';
   var EXPANDED_STATE = 'globalBannerIsExpanded';
 
-  var _dom = atomicHelpers.checkDom( element, BASE_CLASS, 'GlobalBanner' );
+  var _dom = atomicHelpers.checkDom( element, BASE_CLASS );
   var _expandable;
 
   /**